Renamed *.r2dbc.repository to *r2dbc.repository.reactive to match the package name for reactive repository stuff in commons.
Also reenabled the DependencyTest which now longer fails.
Adapted the other test to check the new packages as well.
