+++ src/main/java/org/springframework/data/r2dbc/repository/reactive/package-info.java
@@ -2,6 +2,6 @@
  * R2DBC-specific repository implementation.
  */
 @NonNullApi
-package org.springframework.data.r2dbc.repository;
+package org.springframework.data.r2dbc.repository.reactive;
 
 import org.springframework.lang.NonNullApi;
