+++ src/main/java/org/springframework/data/r2dbc/repository/reactive/support/SimpleR2dbcRepository.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.r2dbc.repository.support;
+package org.springframework.data.r2dbc.repository.reactive.support;
 
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
