+++ src/main/java/org/springframework/data/r2dbc/repository/reactive/support/package-info.java
@@ -2,6 +2,6 @@
  * Support infrastructure for query derivation of R2DBC-specific repositories.
  */
 @NonNullApi
-package org.springframework.data.r2dbc.repository.support;
+package org.springframework.data.r2dbc.repository.reactive.support;
 
 import org.springframework.lang.NonNullApi;
