+++ src/test/java/org/springframework/data/r2dbc/repository/reactive/support/R2dbcRepositoryFactoryUnitTests.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.r2dbc.repository.support;
+package org.springframework.data.r2dbc.repository.reactive.support;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.Mockito.*;
